+++ Changelog.md
@@ -10,6 +10,7 @@
 * Remove usage of Linux `file` command in favor of `ruby-filemagick` gem. Better cross-platform support
 * Use ruby method to create tmp folder instead of calling Linux methods
 * Add method `OoxmlParser::Parser.parse` to parse any OOXML format with auto recognition
+* Add storing color scheme name in color class
 
 ### Fixes
 * Fix misplaced `dxa` and `emu` units of measurements and also fix calculation `dxa` unit